+++ FoodRu_MicroService/src/main/java/org/example/foodru_microservice/repository/MenuRepository.java
@@ -3,5 +3,5 @@ package org.example.foodru_microservice.repository;
 import org.example.foodru_microservice.model.entity.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuRepository extends JpaRepository<Menu, Integer> {
+public interface MenuRepository extends JpaRepository<Menu, Long> {
 }
